AIPCC-2016: Add ignorePrAuthor as tmp workaround
We migrated renovate to a new bot account. And we've set the correct
gitAuthor in the renovate config. This means that the old MRs filed by
the platform-engineering-bot now aren't being rebased as Renovate
doesn't recognize the author.
Enabling ignorePrAuthor we allow Renovate to recreate MRs that were
originally created by the old platform-engineering-bot author.

This is a temporary workaround until all open MRs from
platform-engineering-bot are closed. Once those are resolved, this
setting should be removed to restore proper MR author checking.

Signed-off-by: Giulia Naponiello <gnaponie@redhat.com>